requirements.txt
pytest~=8.3.4
playwright~=1.50.0
requests~=2.32.3
jsonschema~=4.23.0
allure-python-commons~=2.13.5